# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nexus Command API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents/route:
    post:
      operationId: routeTask
      tags: [agents]
      summary: Route a task to the best agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskRouteInput"
      responses:
        "200":
          description: Routing result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskRouteResult"
...
  /projects/{id}/tasks:
    get:
      operationId: listProjectTasks
      tags: [tasks]
      summary: List tasks for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
components:
  schemas:
...
    Agent:
      type: object
      required: [id, name, description, category, reliabilityScore, costScore, status, supportedModels, strengths, useCases]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        reliabilityScore:
          type: number
        costScore:
          type: number
        status:
          type: string
        supportedModels:
          type: array
          items:
            type: string
        strengths:
          type: array
          items:
            type: string
        useCases:
          type: array
          items:
            type: string
        lastUpdated:
          type: string
        permissions:
          type: array
          items:
            type: string
        tasksCompleted:
          type: integer
        avgResponseTime:
          type: number
    TaskRouteInput:
      type: object
      required: [task]
      properties:
        task:
          type: string
    TaskRouteResult:
      type: object
      required: [primaryAgent, secondaryAgent, estimatedCost, confidenceScore, reasoning]
      properties:
        primaryAgent:
          $ref: "#/components/schemas/Agent"
        secondaryAgent:
          $ref: "#/components/schemas/Agent"
        estimatedCost:
          type: number
        confidenceScore:
          type: number
        reasoning:
          type: string
...
    Task:
      type: object
      required: [id, title, status, priority, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        projectId:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        assignee:
          type: string
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
        aiRecommendation:
          type: string
    TaskInput:
      type: object
      required: [title, status, priority]
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        projectId:
          type: ["integer", "null"]
        assignee:
          type: string
        dueDate:
          type: ["string", "null"]
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
        priority:
          type: string
        dueDate:
          type: ["string", "null"]
...