# lib/api-zod/src/generated/types/taskRouteResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexus Command API
 * OpenAPI spec version: 0.1.0
 */
import type { Agent } from './agent';

export interface TaskRouteResult {
  primaryAgent: Agent;
  secondaryAgent: Agent;
  estimatedCost: number;
  confidenceScore: number;
  reasoning: string;
}